requirements.txt
fastapi
python-dotenv
pydantic
pydantic[email]
psycopg2-binary
streamlit
pyttsx3
langchain-core
langchain-huggingface
langchain-google-genai
huggingface-hub
google-genai
google
uvicorn
gunicorn